# pkg/interfaces/router/swagger/ui/docs/swagger.yaml
components:
  schemas:
    response.CsrfToken:
      properties:
        csrf_token:
          type: string
      type: object
    response.EnabledSocialLogin:
      properties:
        socials:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    response.Error:
      properties:
        description:
          type: string
        field:
          type: string
        help:
          type: string
        message:
          type: string
      type: object
    response.Errors:
      properties:
        errors:
          items:
            $ref: '#/components/schemas/response.Error'
          type: array
          uniqueItems: false
      type: object
    response.User:
      properties:
        email:
          type: string
        email_verified:
          type: boolean
        picture:
          type: string
        profile:
          type: string
        sub:
          type: string
        username:
          type: string
      type: object
    response.WebUIManifest:
      properties:
        header_name:
          type: string
        token_base:
          type: boolean
      type: object
externalDocs:
  description: ""
  url: ""
info:
  description: エンドポイントを保護します
  license:
    name: nozomi.nishinohara
  title: Web UI
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/{provider}/authorize:
    get:
      operationId: SocialLoginAuthorize
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "302":
          content:
            application/json:
              schema:
                type: ""
          description: Found
          headers:
            Location:
              schema:
                type: string
      summary: 認証開始エンドポイント
      tags:
      - UI
  /auth/{provider}/callback:
    get:
      operationId: SocialLoginCallback
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "302":
          content:
            application/json:
              schema:
                type: ""
          description: Found
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Errors'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Errors'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Errors'
          description: Internal Server Error
      summary: コールバック
      tags:
      - UI
  /auth/enabled:
    get:
      operationId: EnabledSocialLoginList
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.EnabledSocialLogin'
          description: OK
      summary: 有効なソーシャルログイン
      tags:
      - UI
  /auth/logout:
    get:
      operationId: SocialLogout
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "302":
          content:
            application/json:
              schema:
                type: ""
          description: Found
          headers:
            Location:
              schema:
                type: string
      summary: ログアウト
      tags:
      - UI
  /auth/userinfo:
    get:
      operationId: SocialLoginUserInfo
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.User'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Errors'
          description: Unauthorized
      summary: ログイン済みのユーザー情報取得
      tags:
      - UI
  /csrf_token:
    get:
      operationId: GetCSRFToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.CsrfToken'
          description: OK
      summary: CSRFトークンの生成
      tags:
      - UI
  /manifest:
    get:
      operationId: UIManifest
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.WebUIManifest'
          description: OK
      summary: Manifestの取得
      tags:
      - UI
servers:
- url: /api